Provide BasePluginArgs base class.
BasePluginArgs provides generic implementations of __str__ and supportedNodeosArgs functions for use in each *PluginArgs classes.

Update generation script to no longer generate base class functions, but to inherit from base class.
